Optimize and cleanup
Optimize:
- Increase L1 cache size.
- Stream directly in global memory. Using shared memory showed no
improvement of performance, even in 2d (bad coalescing ?)
—> Performances increased from 119 to 192 MLUPS on a Tesla M2090 (with
NX=500 NY=300 NZ=100).

Cleanup:
- Remove unused right wall grid and block sizes.

Notes:
- Without the « optimized » makefile target, performance drop from 192
to 178 MLUPS.
- The optimal SHARED_MEM_SIZE seems to vary. On a GeForce GT 750M, 0 is
optimal, which seems logic as we do not use shared memory anymore. But,
on a Tesla M2090, performance starts to notably drop (by 10% to 50%)
when the following condition is not met: 4096 < SHARED_MEM_SIZE < 5377.
Reason is yet unknown.